+++ packages/atomic/src/components/atomic-search-interface/atomic-search-interface.tsx
@@ -107,7 +107,20 @@ export class AtomicSearchInterface {
 
   @Watch('language')
   public updateLanguage() {
-    this.i18n.changeLanguage(this.language);
+    new Backend(this.i18n.services, this.i18nBackendOptions).read(
+      this.language,
+      'translation',
+      (_, data) => {
+        this.i18n.addResourceBundle(
+          this.language,
+          'translation',
+          data,
+          true,
+          false
+        );
+        this.i18n.changeLanguage(this.language);
+      }
+    );
   }
 
   public disconnectedCallback() {
@@ -193,9 +206,7 @@ export class AtomicSearchInterface {
       debug: this.logLevel === 'debug',
       lng: this.language,
       fallbackLng: ['en'],
-      backend: {
-        loadPath: `${getAssetPath('./lang/')}{{lng}}.json`,
-      } as BackendOptions,
+      backend: this.i18nBackendOptions,
     });
   }
 
@@ -262,5 +273,11 @@ export class AtomicSearchInterface {
       ),
       <slot></slot>,
     ];
+  }
+
+  private get i18nBackendOptions(): BackendOptions {
+    return {
+      loadPath: `${getAssetPath('./lang/')}{{lng}}.json`,
+    };
   }
 }