+++ packages/headless/src/features/result-preview/result-preview-state.ts
@@ -1,11 +1,13 @@
 export interface ResultPreviewState {
   uniqueId: string;
   content: string;
+  isLoading: boolean;
 }
 
 export function getResultPreviewInitialState(): ResultPreviewState {
   return {
     uniqueId: '',
     content: '',
+    isLoading: false,
   };
 }